requirements.txt
fastapi[standard]
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
